+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Changed
+
+- cosmwasm-vm: Read `Region` from Wasm memory as bytes and convert to `Region`
+  afterwards ([#2005])
+
+[#2005]: https://github.com/CosmWasm/cosmwasm/pull/2005
+
 ## [1.5.2] - 2024-01-15
 
 ### Fixed