+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [1.5.3]
+
 ### Changed
 
 - cosmwasm-vm: Read `Region` from Wasm memory as bytes and convert to `Region`
@@ -619,7 +621,8 @@ and this project adheres to
 The CHANGELOG for versions before 1.0.0 was moved to
 [CHANGELOG-pre1.0.0.md](./CHANGELOG-pre1.0.0.md).
 
-[unreleased]: https://github.com/CosmWasm/cosmwasm/compare/v1.5.2...HEAD
+[unreleased]: https://github.com/CosmWasm/cosmwasm/compare/v1.5.3...HEAD
+[1.5.3]: https://github.com/CosmWasm/cosmwasm/compare/v1.5.2...v1.5.3
 [1.5.2]: https://github.com/CosmWasm/cosmwasm/compare/v1.5.1...v1.5.2
 [1.5.1]: https://github.com/CosmWasm/cosmwasm/compare/v1.5.0...v1.5.1
 [1.5.0]: https://github.com/CosmWasm/cosmwasm/compare/v1.4.1...v1.5.0